Set page searchParam to enable scroll restoration on back navigation
Update the URL to reflect the current page, so that the user can navigate back to this page through history and still have all list items loaded and rendered that were there before.

Refreshing the page will still blow the whole data cache away. We'd need to add some server-side preloading of subsequent pages to fix that.